Вынес страницы в соответствующие модули
